json-format: rename to dash instead of underscore
Maintain naming consistency by renaming json_* to json-*
